Move classes into their own files.
