+++ src/StableDiffusionSD15.py
@@ -1,0 +1,37 @@
+from src.StableDiffusionBase import StableDiffusionBase
+from diffusers import StableDiffusionPipeline
+import torch
+
+
+class StableDiffusionSD15(StableDiffusionBase):
+    def load_pipeline(self):
+        if self.model_id.endswith('.safetensors') or self.model_id.endswith('.ckpt'):
+            pipe = StableDiffusionPipeline.from_single_file(
+                self.model_id,
+                torch_dtype = torch.float16,
+                variant = "fp16",
+                load_safety_checker = False)
+        else:
+            pipe = StableDiffusionPipeline.from_pretrained(
+                self.model_id,
+                torch_dtype = torch.float16,
+                variant = "fp16",
+                safety_checker = None,
+                requires_safety_checker = False)
+        return pipe
+
+    def get_tiny_vae_model_id(self):
+        return 'madebyollin/taesd'
+
+    def get_lcm_adapter_id(self):
+        return "latent-consistency/lcm-lora-sdv1-5"
+    
+    def get_filename_prefix(self):
+        return "sd15-"
+    
+    def get_tokenizers(self):
+        return [self.pipe.tokenizer]
+    
+    def get_text_encoders(self):
+        return [self.pipe.text_encoder]
+
